Implementacion StorageManager y comienzo de MqttManager

+++ src/StorageManager.h
@@ -28,16 +28,49 @@ class StorageManager {
      */
     void createDefaultConfig();  
 
-    void saveMeasure(unsigned long timestamp, int value); 
+    /**
+     * @brief Guarda la medida en /log.csv
+     * 
+     * @param timestamp identifica cuando se tomo la medida 
+     * @param value valor de la medida
+     */
+    void saveMeasure(unsigned long timestamp, double value); 
     
+    /**
+     * @brief Revisa si /log.csv contiene informacion, si contiene informacion es que esta pendiente
+     * 
+     * @return true tiene datos pendientes 
+     * @return false no tiene datos pendientes
+     */
     bool hasPendingData(); 
 
+    /**
+     * @brief Si hay infomracion pendiente en  "/log.csv" lo renombre a "/temp.csv"
+     * 
+     * @return true si hay informacion y se renombbre a /temp.csv
+     * @return false si no hay informacion
+     */
     bool PrepareDataForUpload(); 
     
+    /**
+     * @brief Abre el archivo /temp,csv
+     * 
+     * @return File 
+     */
     File openTempFile(); 
-    
+
+    /**
+     * @brief Get the Next Line 
+     * 
+     * @param file Archivo que se esta leyendo
+     * @return String cadena que devuelve
+     */
     String getNextLine(File &file); 
     
+    /**
+     * @brief Borra el archivo /temp.csv
+     * 
+     */
     void finishDataUpload(); 
 
 };
